Moved data path definitions into __init__.py
