--- src/pulsar_spectra/load_data.py
@@ -1,7 +1,0 @@
-import os
-
-CONFIG_DIR = os.path.join(os.path.dirname(__file__), "configs")
-
-DEFAULT_PLOTTING_CONFIG = os.path.join(os.path.dirname(__file__), "configs/plotting_config.yaml")
-
-DEFAULT_MARKER_CSV = os.path.join(os.path.dirname(__file__), "configs/default_markers.csv")
