+++ src/pulsar_spectra/scripts/build_plotting_config.py
@@ -14,7 +14,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import yaml
 
-from pulsar_spectra.load_data import DEFAULT_MARKER_CSV
+from pulsar_spectra import DEFAULT_MARKER_CSV
 
 # [colour, description]
 # Standard matplotlib colours
